♻️refactor: rename carousel
