+++ app/area/tokyo_chiba/page.tsx
@@ -1,11 +1,11 @@
-import { PlanCarousel } from "@/app/components/layout/carousel"
+import { ScheduleCarousel } from "@/app/components/layout/carousel"
 import { Heading } from "@/app/components/layout/heading"
 import { Video } from "@/app/components/media/video"
+import type { Schedule } from "@/app/interfaces/schedule"
 import {
   AREA,
   TOKYO_CHIBA,
   TOKYO_CHIBA_LINE_LINK,
-  TokyoChibaCarouselItems,
   TokyoChibaEvents,
 } from "@/app/lib/constant"
 import { ArrowRightIcon } from "@heroicons/react/24/solid"
@@ -14,6 +14,33 @@ import Link from "next/link"
 import type React from "react"
 
 export default function Chiba(): React.JSX.Element {
+  const schedules: Schedule[] = [
+    {
+      alt: "ロボット制作",
+      src: "/202311/sandankyo/brothers.avif",
+      color: "bg-teal-400",
+      title: "午前：ロボット制作\n午後：自然の循環",
+      date: "2024年9月14日(土) 10:00~17:00",
+      tags: ["ロボット制作", "自然学習"],
+    },
+    {
+      alt: "謎解き体験！",
+      src: "/content/rays-blog_girl.avif",
+      color: "bg-info",
+      title: "午前：プログラミング体験\n午後：謎解き体験！",
+      date: "2024年9月21日(土) 10:00~17:00",
+      tags: ["プログラミング", "情報教育"],
+    },
+    {
+      alt: "ロボサバ大会",
+      src: "/202307/final_stage/course_challenge.avif",
+      color: "bg-rose-400",
+      title: "午前：ロボサバ大会\n午後：結婚式体験",
+      date: "2024年9月22日(日) 10:00~17:00",
+      tags: ["ロボサバ", "結婚式体験"],
+    },
+  ] as const
+
   return (
     <>
       <Heading navigation={AREA} content={TOKYO_CHIBA} />
@@ -74,7 +101,7 @@ export default function Chiba(): React.JSX.Element {
       </section>
       <section className="grid gap-4">
         <h2 className="font-bold font-zenMaruGothic text-3xl">スケジュール</h2>
-        <PlanCarousel carouselItems={TokyoChibaCarouselItems} />
+        <ScheduleCarousel schedules={schedules} />
       </section>
       <Video src="https://dk75m1tgsot44.cloudfront.net/movie/202312" />
     </>
