--- app/interfaces/carousel.ts
@@ -1,28 +1,0 @@
-export type Carousel = {
-  alt: string
-  src: string
-}
-
-export type EventCarousel = {
-  alt: string
-  src: string
-  category: string
-  color: string
-  title: string
-  date: string
-  tags: string[]
-}
-
-export type ReviewCarousel = {
-  description: string
-  areaAndUser: string
-}
-
-export type ScheduleCarousel = {
-  alt: string
-  src: string
-  color: string
-  title: string
-  date: string
-  tags: string[]
-}
