+++ app/interfaces/picture.ts
@@ -1,4 +1,4 @@
-export type Panel = {
+export type Picture = {
   alt: string
   src: string
 }
