+++ app/interfaces/review.ts
@@ -1,0 +1,4 @@
+export type Review = {
+  description: string
+  areaAndUser: string
+}
