+++ app/interfaces/schedule.ts
@@ -1,0 +1,8 @@
+export type Schedule = {
+  alt: string
+  src: string
+  color: string
+  title: string
+  date: string
+  tags: string[]
+}
